Renamed package logging to log
