+++ Open-Master/src/main/java/de/tammo/cloud/master/Master.java
@@ -10,8 +10,8 @@ import de.tammo.cloud.command.CommandProviderService;
 import de.tammo.cloud.config.DocumentProviderService;
 import de.tammo.cloud.core.CloudApplication;
 import de.tammo.cloud.core.file.FileUtils;
-import de.tammo.cloud.core.logging.LogLevel;
-import de.tammo.cloud.core.logging.Logger;
+import de.tammo.cloud.core.log.LogLevel;
+import de.tammo.cloud.core.log.Logger;
 import de.tammo.cloud.core.service.ServiceProvider;
 import de.tammo.cloud.master.components.ComponentsProviderService;
 import de.tammo.cloud.master.config.configuration.Configuration;
