+++ Open-Master/src/main/java/de/tammo/cloud/master/network/handler/PacketHandler.java
@@ -6,7 +6,7 @@
 
 package de.tammo.cloud.master.network.handler;
 
-import de.tammo.cloud.core.logging.Logger;
+import de.tammo.cloud.core.log.Logger;
 import de.tammo.cloud.core.service.ServiceProvider;
 import de.tammo.cloud.master.network.NetworkProviderService;
 import de.tammo.cloud.master.network.packets.in.WrapperKeyInPacket;
