+++ Open-Modules/Open-Command/src/main/java/de/tammo/cloud/command/CommandProviderService.java
@@ -7,7 +7,7 @@
 package de.tammo.cloud.command;
 
 import com.google.common.reflect.ClassPath;
-import de.tammo.cloud.core.logging.Logger;
+import de.tammo.cloud.core.log.Logger;
 import de.tammo.cloud.core.service.Service;
 import lombok.Getter;
