+++ Open-Wrapper/src/main/java/de/tammo/cloud/wrapper/commands/HelpCommand.java
@@ -8,7 +8,7 @@ package de.tammo.cloud.wrapper.commands;
 
 import de.tammo.cloud.command.Command;
 import de.tammo.cloud.command.CommandProviderService;
-import de.tammo.cloud.core.logging.Logger;
+import de.tammo.cloud.core.log.Logger;
 import de.tammo.cloud.core.service.ServiceProvider;
 
 @Command.CommandInfo(name = "help")
