+++ Open-Wrapper/src/main/java/de/tammo/cloud/wrapper/components/proxy/ProxyServer.java
@@ -7,7 +7,7 @@
 package de.tammo.cloud.wrapper.components.proxy;
 
 import de.tammo.cloud.core.file.FileUtils;
-import de.tammo.cloud.core.logging.Logger;
+import de.tammo.cloud.core.log.Logger;
 import de.tammo.cloud.core.service.ServiceProvider;
 import de.tammo.cloud.wrapper.Wrapper;
 import de.tammo.cloud.wrapper.components.ServerComponent;
