+++ Open-Wrapper/src/main/java/de/tammo/cloud/wrapper/components/server/GameServer.java
@@ -7,7 +7,7 @@
 package de.tammo.cloud.wrapper.components.server;
 
 import de.tammo.cloud.core.file.FileUtils;
-import de.tammo.cloud.core.logging.Logger;
+import de.tammo.cloud.core.log.Logger;
 import de.tammo.cloud.core.service.ServiceProvider;
 import de.tammo.cloud.wrapper.Wrapper;
 import de.tammo.cloud.wrapper.components.ServerComponent;
